Longest Palindromic Substring
You are given a string s. Your task is to find and return the longest palindromic substring within the given string. A palindrome is a string that reads the same forwards and backwards.

Input:
A string s of length n. The length of the string satisfies 1≤n≤1000.

Output:
Return the longest substring of s that is a palindrome. If there are multiple solutions, return the first one that occurs.