Find the Shortest Path in an Unweighted Graph
You are given an unweighted graph represented by an adjacency list. Your task is to find the shortest path (in terms of the number of edges) between two given nodes in the graph.

Input:
An integer V representing the number of vertices in the graph.
A list of edges, where each edge connects two vertices of the graph.
Two integers, start and end, representing the source and destination nodes respectively.

Output:
Return the shortest path length (number of edges) from start to end. If there is no path, return -1.